+++ NEWS.md
@@ -15,6 +15,7 @@
 
 ## New features, bug fixes, improvements
 
+* Add exception for `box::use()` declarations to infix spaces linter (#1087, @klmr)
 * Writes comments to GitHub repo when running in Jenkins CI (#488, @fdlk)
 * Updated R CMD GitHub Actions workflow to check for R 3.6 on Ubuntu, instead of R 3.3, and for R 4.0 on Windows, instead of R 3.6 (#803, @ dragosmg)
 * Added a secondary, more restrictive lint workflow - `lint-changed-files` - for newly written / modified code (#641, @dragosmg)